+++ tests/test_search_itp_gecodeerdelithologie.py
@@ -1,11 +1,8 @@
 """Module grouping tests for the interpretaties search module."""
-import numpy as np
 import pandas as pd
-import pytest
 from owslib.fes import PropertyIsEqualTo
 from pandas import DataFrame
 
-import pydov
 from pydov.search.interpretaties import GecodeerdeLithologieSearch
 from pydov.types.interpretaties import GecodeerdeLithologie
 from tests.abstract import AbstractTestSearch
@@ -180,8 +177,7 @@ class TestGecodeerdeLithologieSearch(AbstractTestSearch):
             Monkeypatch the call to get the remote XML data.
 
         """
-        df = self.get_search_object().search(
-            query=self.get_valid_query_single())
+        self.get_search_object().search(query=self.get_valid_query_single())
 
     def test_search_customreturnfields(self, mp_get_schema,
                                        mp_remote_describefeaturetype,
