+++ tests/test_types_itp_gecodeerdelithologie.py
@@ -10,6 +10,7 @@ from tests.test_search_itp_gecodeerdelithologie import (
 class TestGecodeerdeLithologie(AbstractTestTypes):
     """Class grouping tests for the
     pydov.types.interpretaties.GecodeerdeLithologie class."""
+
     def get_type(self):
         """Get the class reference for this datatype.
 
@@ -58,9 +59,11 @@ class TestGecodeerdeLithologie(AbstractTestTypes):
                 'betrouwbaarheid_interpretatie', 'x', 'y',
                 'diepte_laag_van', 'diepte_laag_tot',
                 'hoofdnaam1_grondsoort', 'hoofdnaam2_grondsoort',
-                'bijmenging1_plaatselijk', 'bijmenging1_hoeveelheid', 'bijmenging1_grondsoort',
-                'bijmenging2_plaatselijk', 'bijmenging2_hoeveelheid', 'bijmenging2_grondsoort',
-                'bijmenging3_plaatselijk', 'bijmenging3_hoeveelheid', 'bijmenging3_grondsoort',]
+                'bijmenging1_plaatselijk', 'bijmenging1_hoeveelheid',
+                'bijmenging1_grondsoort', 'bijmenging2_plaatselijk',
+                'bijmenging2_hoeveelheid', 'bijmenging2_grondsoort',
+                'bijmenging3_plaatselijk', 'bijmenging3_hoeveelheid',
+                'bijmenging3_grondsoort']
 
     def get_field_names_subtypes(self):
         """Get the field names of this type that originate from subtypes only.
@@ -73,9 +76,11 @@ class TestGecodeerdeLithologie(AbstractTestTypes):
         """
         return ['diepte_laag_van', 'diepte_laag_tot',
                 'hoofdnaam1_grondsoort', 'hoofdnaam2_grondsoort',
-                'bijmenging1_plaatselijk', 'bijmenging1_hoeveelheid', 'bijmenging1_grondsoort',
-                'bijmenging2_plaatselijk', 'bijmenging2_hoeveelheid', 'bijmenging2_grondsoort',
-                'bijmenging3_plaatselijk', 'bijmenging3_hoeveelheid', 'bijmenging3_grondsoort',]
+                'bijmenging1_plaatselijk', 'bijmenging1_hoeveelheid',
+                'bijmenging1_grondsoort', 'bijmenging2_plaatselijk',
+                'bijmenging2_hoeveelheid', 'bijmenging2_grondsoort',
+                'bijmenging3_plaatselijk', 'bijmenging3_hoeveelheid',
+                'bijmenging3_grondsoort']
 
     def get_field_names_nosubtypes(self):
         """Get the field names for this type, without including fields from
